# backend/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Code Review Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
  reviewCount: number;
  projectCount: number;
}